Move AutoApplySessionInterceptor to extensions package for issue #408
